src/ilirsubashialumin/DBSconnect.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package ilirsubashialumin;

import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.SQLException;
import java.sql.Statement;
import javax.swing.JOptionPane;

/**
 *
 * @author bokae
 */
public class DBSconnect {
    public static Connection connect()
            
    {
    Connection con=null;
    try{
            Class.forName("com.mysql.jdbc.Driver");
            con = DriverManager.getConnection("jdbc:mysql://localhost/projekti punes", "root", "");
        
        }
        catch(SQLException | ClassNotFoundException e){
            System.out.println("inter.DBConnect.connect()");
            JOptionPane.showMessageDialog(null, e);
        }
    
    return con;
    }
    
}
